backend/requirements.txt
fastapi
uvicorn
google-genai
langgraph
pypdf
python-dotenv
chromadb
python-multipart